create _EditGlosses component
