+++ src/routes/[dictionaryId]/_EditString.svelte
